WIP: add reexec logic to hyperv
