+++ pkg/machine/hyperv/hutil.go
@@ -3,12 +3,17 @@
 package hyperv
 
 import (
+	"fmt"
+
+	"github.com/containers/podman/v5/pkg/machine/define"
 	"github.com/sirupsen/logrus"
-	"golang.org/x/sys/windows"
+	win "golang.org/x/sys/windows"
+
+	"github.com/containers/podman/v5/pkg/machine/windows"
 )
 
 func HasHyperVAdminRights() bool {
-	sid, err := windows.CreateWellKnownSid(windows.WinBuiltinHyperVAdminsSid)
+	sid, err := win.CreateWellKnownSid(win.WinBuiltinHyperVAdminsSid)
 	if err != nil {
 		return false
 	}
@@ -18,7 +23,7 @@ func HasHyperVAdminRights() bool {
 	//  token of the calling thread. If the thread is not impersonating,
 	//  the function duplicates the thread's primary token to create an
 	//  impersonation token."
-	token := windows.Token(0)
+	token := win.Token(0)
 	member, err := token.IsMember(sid)
 
 	if err != nil {
@@ -27,4 +32,12 @@ func HasHyperVAdminRights() bool {
 	}
 
 	return member
+}
+
+func LaunchHyperVElevated() error {
+	fmt.Println("Launching HyperV as Admin to add user to hyperv Admin and registry...")
+	if err := windows.LaunchElevate("Add user to HyperV Admin group", "Could not %s. See previous output for any potential failure details."); err != nil {
+		return define.ErrInitRelaunchAttempt
+	}
+	return nil
 }
